% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomp-list.R
\name{Decomp}
\alias{Decomp}
\title{Generate Simulated Data for Oaxaca Decomposition}
\usage{
Decomp(results, name = "Decomposition", decomp = "threefold")
}
\arguments{
\item{results}{a oaxaca object}

\item{name}{a customisable output name}

\item{decomp}{either 'threefold' or 'twofold'}
}
\value{
A list
}
\description{
Generate Simulated Data for Oaxaca Decomposition
}
\examples{
dff = oaxaca_data()
output = oaxaca(Y ~ X | group, data = dff$dataframe)
Decomp(output)
Decomp(output, decomp = 'twofold')
}
