% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomp_InfoBlinderOax.R
\name{Decomp_info_OaxacaBlinder}
\alias{Decomp_info_OaxacaBlinder}
\title{Generate Simulated Data for Oaxaca Decomposition}
\usage{
Decomp_info_OaxacaBlinder(results, full = F, SE = F)
}
\arguments{
\item{results}{a oaxaca object}
}
\value{
A list
}
\description{
Generate Simulated Data for Oaxaca Decomposition
}
\examples{
dff = oaxaca_data()
output = oaxaca(Y ~ X | group, data = dff$dataframe)
Decomp_info_OB(output)
}
