% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomp_regression.R
\name{Decomp_Regression}
\alias{Decomp_Regression}
\title{Generate Simulated Data for Oaxaca Decomposition}
\usage{
Decomp_Regression(oax_obj)
}
\arguments{
\item{oax_obj}{a oaxaca object}
}
\value{
A list
}
\description{
Generate Simulated Data for Oaxaca Decomposition
}
\examples{
dff = oaxaca_data()
output = oaxaca(Y ~ X | group, data = dff$dataframe)
Decomp_Regression(output)
}
