% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomp_simple.R
\name{Decomp_simple}
\alias{Decomp_simple}
\title{Generate Simulated Data for Oaxaca Decomposition}
\usage{
Decomp_simple(results)
}
\arguments{
\item{results}{a oaxaca object}
}
\value{
A list
}
\description{
Generate Simulated Data for Oaxaca Decomposition
}
\examples{
dff = oaxaca_data()
output = oaxaca(Y ~ X | group, data = dff$dataframe)
Decomp_simple(output)
}
