% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Decomp_simdata.R
\name{oaxaca_data}
\alias{oaxaca_data}
\title{Generate Simulated Data for Oaxaca Decomposition}
\usage{
oaxaca_data(
  interceptA_avr = 100,
  interceptB_avr = 105,
  betaA = 5,
  betaB = 8,
  xA_avr = 10,
  xA_sd = 5,
  xB_avr = 10,
  xB_sd = 5,
  errorA_avr = 0,
  errorA_sd = 1,
  errorB_avr = 0,
  errorB_sd = 1,
  N = 1000
)
}
\arguments{
\item{interceptA_avr}{intercept of group A}

\item{interceptB_avr}{intercept of group B}

\item{betaA}{Coefficient size of group A}

\item{betaB}{Coefficient size of group B}

\item{xA_avr}{Group A characteristics}

\item{xA_sd}{Variation in group A characteristics (SD)}

\item{xB_avr}{Group B characteristics}

\item{xB_sd}{Variation in group B characteristics (SD)}
}
\value{
A list
}
\description{
Generate Simulated Data for Oaxaca Decomposition
}
\examples{
oaxaca_data(betaA = 5, betaB = 5, xA_avr = 5, xB_avr = 20)
}
